doc: Add Chan et al. (2024) to bibliography
Adds the paper ""Is Hate Lost in Translation?": Evaluation of Multilingual LGBTQIA+ Hate Speech Detection" by Chan et al. (2024) to the Queer NLP bibliography.

This paper is added to the "Toxicity and Hate/Hope Speech", "LLMs", and "Translation" sections of the README.md file and the corresponding BibTeX entry is added to the refs.bib file.